fix:clear platform tool scan problems

+++ adapter/khdf/linux/platform/mipi_dsi/mipi_tx_hi35xx.c
@@ -532,7 +532,7 @@ static void SetLaneConfig(const short laneId[], int len)
             num++;
         }
     }
-    g_mipiTxRegsVa->PHY_IF_CFG.u32 = num - 1;
+    g_mipiTxRegsVa->PHY_IF_CFG.u32 = (unsigned int)(num - 1);
 }
 
 static void MipiTxDrvSetClkMgrCfg(void)
